fix(vscode): install settings.json even if it doesn't exist

+++ vscode/vscode.sh
@@ -6,17 +6,14 @@ target="$HOME/Library/Application Support/Code/User/$settings"
 
 source "$vscode_path/vscode_extensions.sh"
 
-
-if [[ -e $target ]]; then
-  if [[ ! -L $target ]]; then
-    echo "WARNING: original '$target' will be removed"
-    rm -f "$target"
-  fi
-
-  echo "[VSCode]: symlink user's $vscode_path/$settings $target"
-
-  ln -sf "$vscode_path/$settings" "$target"
+if [[ -e $target && ! -L $target ]]; then
+  echo "WARNING: original '$target' will be removed"
+  rm -f "$target"
 fi
+
+echo "[VSCode]: symlink user's $vscode_path/$settings $target"
+
+ln -sf "$vscode_path/$settings" "$target"
 
 if [ -x $(which code) ]; then
   for (( i=0; i<${#extensions[*]}; i++ ))
